requirements.txt
Flask
numpy
pandas
tensorflow-cpu
joblib
scikit-learn
